requirements.txt
numpy>=1.22
pillow>=9.0
scipy>=1.9
matplotlib>=3.6
